+++ uni/lib/controller/load_info.dart
@@ -2,17 +2,15 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:connectivity_plus/connectivity_plus.dart';
-import 'package:redux/redux.dart';
+import 'package:logger/logger.dart';
 import 'package:tuple/tuple.dart';
 import 'package:uni/controller/local_storage/app_shared_preferences.dart';
 import 'package:uni/controller/local_storage/file_offline_storage.dart';
 import 'package:uni/controller/parsers/parser_exams.dart';
-import 'package:uni/model/app_state.dart';
-import 'package:uni/redux/action_creators.dart';
-import 'package:uni/redux/actions.dart';
-import 'package:uni/redux/refresh_items_action.dart';
+import 'package:uni/model/entities/session.dart';
+import 'package:uni/model/providers/state_providers.dart';
 
-Future loadReloginInfo(Store<AppState> store) async {
+Future loadReloginInfo(StateProviders stateProviders) async {
   final Tuple2<String, String> userPersistentCredentials =
       await AppSharedPreferences.getPersistentUserInfo();
   final String userName = userPersistentCredentials.item1;
@@ -21,25 +19,23 @@ Future loadReloginInfo(Store<AppState> store) async {
 
   if (userName != '' && password != '') {
     final action = Completer();
-    store.dispatch(reLogin(userName, password, faculties, action: action));
+    stateProviders.sessionProvider
+        .reLogin(userName, password, faculties, stateProviders, action: action);
     return action.future;
   }
   return Future.error('No credentials stored');
 }
 
-Future loadUserInfoToState(store) async {
-  loadLocalUserInfoToState(store);
-  if (await Connectivity().checkConnectivity() != ConnectivityResult.none) {
-    return loadRemoteUserInfoToState(store);
+Future loadRemoteUserInfoToState(StateProviders stateProviders) async {
+  if (await Connectivity().checkConnectivity() == ConnectivityResult.none) {
+    return;
   }
-}
 
-Future loadRemoteUserInfoToState(Store<AppState> store) async {
-  if (store.state.content['session'] == null) {
-    return null;
-  } else if (!store.state.content['session'].authenticated &&
-      store.state.content['session'].persistentSession) {
-    await loadReloginInfo(store);
+  Logger().i('Loading remote info');
+
+  final session = stateProviders.sessionProvider.session;
+  if (!session.authenticated && session.persistentSession) {
+    await loadReloginInfo(stateProviders);
   }
 
   final Completer<void> userInfo = Completer(),
@@ -54,20 +50,28 @@ Future loadRemoteUserInfoToState(Store<AppState> store) async {
       libraryOccupation = Completer(),
       calendar = Completer();
 
-  store.dispatch(getUserInfo(userInfo));
-  store.dispatch(getUserPrintBalance(printBalance));
-  store.dispatch(getUserFees(fees));
-  store.dispatch(getUserBusTrips(trips));
-  store.dispatch(getRestaurantsFromFetcher(restaurants));
-  store.dispatch(getLibraryOccupationFromFetcher(libraryOccupation));
-  store.dispatch(getCalendarFromFetcher(calendar));
+  stateProviders.profileStateProvider.getUserInfo(userInfo, session);
+  stateProviders.busStopProvider.getUserBusTrips(trips);
+  stateProviders.restaurantProvider
+      .getRestaurantsFromFetcher(restaurants, session);
+  stateProviders.calendarProvider.getCalendarFromFetcher(session, calendar);
+  stateProviders.libraryOccupationProvider
+      .getLibraryOccupation(session, libraryOccupation);
 
   final Tuple2<String, String> userPersistentInfo =
       await AppSharedPreferences.getPersistentUserInfo();
   userInfo.future.then((value) {
-    store.dispatch(getUserExams(exams, ParserExams(), userPersistentInfo));
-    store.dispatch(getUserSchedule(schedule, userPersistentInfo));
-    store.dispatch(getCourseUnitsAndCourseAverages(ucs));
+    final profile = stateProviders.profileStateProvider.profile;
+    final currUcs = stateProviders.profileStateProvider.currUcs;
+    stateProviders.examProvider.getUserExams(
+        exams, ParserExams(), userPersistentInfo, profile, session, currUcs);
+    stateProviders.lectureProvider
+        .getUserLectures(schedule, userPersistentInfo, session, profile);
+    stateProviders.profileStateProvider
+        .getCourseUnitsAndCourseAverages(session, ucs);
+    stateProviders.profileStateProvider
+        .getUserPrintBalance(printBalance, session);
+    stateProviders.profileStateProvider.getUserFees(fees, session);
   });
 
   final allRequests = Future.wait([
@@ -83,54 +87,61 @@ Future loadRemoteUserInfoToState(Store<AppState> store) async {
     calendar.future
   ]);
   allRequests.then((futures) {
-    store.dispatch(setLastUserInfoUpdateTimestamp(lastUpdate));
+    stateProviders.lastUserInfoProvider
+        .setLastUserInfoUpdateTimestamp(lastUpdate);
   });
   return lastUpdate.future;
 }
 
-void loadLocalUserInfoToState(store) async {
-  store.dispatch(
-      UpdateFavoriteCards(await AppSharedPreferences.getFavoriteCards()));
-  store.dispatch(SetExamFilter(await AppSharedPreferences.getFilteredExams()));
-  store.dispatch(
-      SetLocationFilter(await AppSharedPreferences.getFilteredLocations()));
-  store.dispatch(SetExamHidden(await AppSharedPreferences.getHiddenExams()));
-  store.dispatch(
-      SetUserFaculties(await AppSharedPreferences.getUserFaculties()));
+void loadLocalUserInfoToState(StateProviders stateProviders,
+    {skipDatabaseLookup = false}) async {
   final Tuple2<String, String> userPersistentInfo =
       await AppSharedPreferences.getPersistentUserInfo();
-  if (userPersistentInfo.item1 != '' && userPersistentInfo.item2 != '') {
-    store.dispatch(updateStateBasedOnLocalProfile());
-    store.dispatch(updateStateBasedOnLocalUserExams());
-    store.dispatch(updateStateBasedOnLocalUserLectures());
-    store.dispatch(updateStateBasedOnLocalUserBusStops());
-    store.dispatch(updateStateBasedOnLocalRefreshTimes());
-    store.dispatch(updateStateBasedOnLocalTime());
-    store.dispatch(updateStateBasedOnLocalCalendar());
-    store.dispatch(updateRestaurantsBasedOnLocalData());
-    store.dispatch(updateStateBasedOnLocalCourseUnits());
-    store.dispatch(SaveProfileStatusAction(RequestStatus.successful));
-    store.dispatch(SetPrintBalanceStatusAction(RequestStatus.successful));
-    store.dispatch(SetFeesStatusAction(RequestStatus.successful));
+
+  Logger().i('Setting up user preferences');
+  stateProviders.favoriteCardsProvider
+      .setFavoriteCards(await AppSharedPreferences.getFavoriteCards());
+  stateProviders.examProvider.setFilteredExams(
+      await AppSharedPreferences.getFilteredExams(), Completer());
+  stateProviders.facultyLocationsProvider.setFilteredLocations(
+      await AppSharedPreferences.getFilteredLocations(), Completer());
+  stateProviders.examProvider
+      .setHiddenExams(await AppSharedPreferences.getHiddenExams(), Completer());
+  stateProviders.userFacultiesProvider
+      .setUserFaculties(await AppSharedPreferences.getUserFaculties());
+
+  if (userPersistentInfo.item1 != '' &&
+      userPersistentInfo.item2 != '' &&
+      !skipDatabaseLookup) {
+    Logger().i('Fetching local info from database');
+    stateProviders.examProvider.updateStateBasedOnLocalUserExams();
+    stateProviders.lectureProvider.updateStateBasedOnLocalUserLectures();
+    stateProviders.examProvider
+        .updateStateBasedOnLocalUserExams(); // TODO ... n está repetido
+    stateProviders.lectureProvider.updateStateBasedOnLocalUserLectures();
+    stateProviders.busStopProvider.updateStateBasedOnLocalUserBusStops();
+    stateProviders.profileStateProvider.updateStateBasedOnLocalProfile();
+    stateProviders.profileStateProvider.updateStateBasedOnLocalRefreshTimes();
+    stateProviders.restaurantProvider.updateStateBasedOnLocalRestaurants();
+    stateProviders.lastUserInfoProvider.updateStateBasedOnLocalTime();
+    stateProviders.calendarProvider.updateStateBasedOnLocalCalendar();
+    stateProviders.profileStateProvider.updateStateBasedOnLocalCourseUnits();
   }
-  final Completer locations = Completer();
-  store.dispatch(getFacultyLocations(locations));
+
+  stateProviders.facultyLocationsProvider.getFacultyLocations(Completer());
 }
 
-Future<void> handleRefresh(store) {
-  final action = RefreshItemsAction();
-  store.dispatch(action);
-  return action.completer.future;
+Future<void> handleRefresh(StateProviders stateProviders) async {
+  await loadRemoteUserInfoToState(stateProviders);
 }
 
-Future<File?> loadProfilePicture(Store<AppState> store,
-    {forceRetrieval = false}) {
-  final String studentNumber = store.state.content['session'].studentNumber;
-  final String faculty = store.state.content['session'].faculties[0];
+Future<File?> loadProfilePicture(Session session, {forceRetrieval = false}) {
+  final String studentNumber = session.studentNumber;
+  final String faculty = session.faculties[0];
   final String url =
       'https://sigarra.up.pt/$faculty/pt/fotografias_service.foto?pct_cod=$studentNumber';
   final Map<String, String> headers = <String, String>{};
-  headers['cookie'] = store.state.content['session'].cookies;
+  headers['cookie'] = session.cookies;
   return loadFileFromStorageOrRetrieveNew('user_profile_picture', url, headers,
       forceRetrieval: forceRetrieval);
 }