+++ uni/lib/controller/local_storage/app_shared_preferences.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:encrypt/encrypt.dart' as encrypt;
 import 'package:flutter/material.dart';
-import 'package:logger/logger.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:tuple/tuple.dart';
 import 'package:uni/model/entities/exam.dart';
@@ -19,6 +18,7 @@ class AppSharedPreferences {
   static const String userFaculties = 'user_faculties';
   static const String termsAndConditions = 'terms_and_conditions';
   static const String areTermsAndConditionsAcceptedKey = 'is_t&c_accepted';
+  static const String tuitionNotificationsToggleKey = "tuition_notification_toogle";
   static const String themeMode = 'theme_mode';
   static const int keyLength = 32;
   static const int ivLength = 16;
@@ -44,9 +44,9 @@ class AppSharedPreferences {
   /// Saves the user's student number, password and faculties.
   static Future savePersistentUserInfo(user, pass, faculties) async {
     final prefs = await SharedPreferences.getInstance();
-    prefs.setString(userNumber, user);
-    prefs.setString(userPw, encode(pass));
-    prefs.setStringList(
+    await prefs.setString(userNumber, user);
+    await prefs.setString(userPw, encode(pass));
+    await prefs.setStringList(
         userFaculties, faculties); // Could be multiple faculties
   }
 
@@ -136,8 +136,6 @@ class AppSharedPreferences {
 
     if (pass != '') {
       pass = decode(pass);
-    } else {
-      Logger().w('User password does not exist in shared preferences.');
     }
 
     return pass;
@@ -238,4 +236,15 @@ class AppSharedPreferences {
     final key = encrypt.Key.fromLength(keyLength);
     return encrypt.Encrypter(encrypt.AES(key));
   }
+
+  static Future<bool> getTuitionNotificationToggle() async{
+    final prefs = await SharedPreferences.getInstance();
+    return prefs.getBool(tuitionNotificationsToggleKey) ?? true;
+  }
+
+  static setTuitionNotificationToggle(bool value) async{
+    final prefs = await SharedPreferences.getInstance();
+    prefs.setBool(tuitionNotificationsToggleKey, value);
+  }
+
 }